refactor: improve tree-shaking for logging code
